basic clickhouse vector store client

+++ src/rag/Embedding.scala
@@ -4,12 +4,12 @@ package rag
 enum Embedding:
   case Index(
       chunk: Chunk,
-      value: Vector[Double],
+      value: Vector[Float],
       documentId: String,
       fragmentIndex: Int
   )
 
   case Query(
       chunk: Chunk,
-      value: Vector[Double]
+      value: Vector[Float]
   )
